favor long format

+++ src/utils/relativeTime.ts
@@ -10,7 +10,7 @@ const units = {
 
 const rtf = new Intl.RelativeTimeFormat('en', {
   numeric: 'auto',
-  style: 'short',
+  style: 'long',
 });
 
 export const getRelativeTime = (date: Date) => {
@@ -22,4 +22,3 @@ export const getRelativeTime = (date: Date) => {
     }
   }
 };
-
